requirements.txt
openai
openai-whisper
langchain
pytube
youtube-transcript-api
tiktoken
redis